エラー修正の為更に戻す

+++ app/controllers/linebot_controller.rb
@@ -1,10 +1,8 @@
 class LinebotController < ApplicationController
   require 'line/bot' 
-
-  # callbackアクションのCSRFトークン認証を無効
+ 
   protect_from_forgery :except => [:callback]
  
-
   def client
     @client ||= Line::Bot::Client.new { |config|
       config.channel_secret = ENV["LINE_CHANNEL_SECRET"]
@@ -12,7 +10,6 @@ class LinebotController < ApplicationController
     }
   end
  
-
   def callback
     body = request.body.read
 
@@ -22,17 +19,30 @@ class LinebotController < ApplicationController
     end
 
     events = client.parse_events_from(body)
+
     events.each { |event|
-      case event
-      when Line::Bot::Event::Message
+    case event
+    when Line::Bot::Event::Message
+      
       case event.type
-        when Line::Bot::Event::MessageType::Text
-          message = {
-            type: 'text',
-              user = User.find_by(uid: userId)
-              user.destroy if user.present?
-            end
+      when Line::Bot::Event::MessageType::Text
+        message = {
+          type: 'text',
+          text: event.message['text']
+        }
+        client.reply_message(event['replyToken'], message)
+        when Line::Bot::Event::MessageType::Follow #友達登録イベント
+          userId = event['source']['userId'] 
+          User.find_or_create_by(uid: userId)
+          when Line::Bot::Event::MessageType::Unfollow#友達削除イベント
+            userId = event['source']['userId']  
+            user = User.find_by(uid: userId)
+            user.destroy if user.present?
+
           end
+        end
+      end
+    end
     }
 
     head :ok
